pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "peph"
version = "0.1.0"
description = "Piecewise exponential proportional hazards (PE-PH) model package"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "scipy>=1.10",
  "statsmodels>=0.14",
  "pyyaml>=6.0",
  "pydantic>=2.0",
  "matplotlib>=3.8"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-cov>=4.1",
]

[project.scripts]
peph = "peph.cli.main:main"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
markers = [
  "slow: slow simulation or recovery tests",
]